Move kafka-queue logic under src/main
